feat: import genetic analysis orders in genetic analyst pallet.
