+++ pallets/genetic-analysts/benchmarking/src/mock.rs
@@ -1,7 +1,10 @@
+use crate as genetic_analysts;
 use frame_support::{parameter_types, PalletId};
+use frame_system as system;
 use pallet_balances::AccountData;
 use scale_info::TypeInfo;
 use sp_core::{Decode, Encode, RuntimeDebug, H256};
+use sp_io::TestExternalities;
 use sp_runtime::{
 	testing::Header,
 	traits::{BlakeTwo256, IdentityLookup},
@@ -9,6 +12,9 @@ use sp_runtime::{
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
 type Block = frame_system::mocking::MockBlock<Test>;
+
+#[derive(Clone, Copy, PartialEq, Eq, Encode, Decode, Default, RuntimeDebug, TypeInfo)]
+pub struct EthereumAddress(pub [u8; 20]);
 
 pub type AccountId = u64;
 
@@ -26,18 +32,18 @@ frame_support::construct_runtime!(
 		GeneticAnalystQualifications: genetic_analyst_qualifications::{Pallet, Call, Storage, Event<T>},
 		GeneticAnalysis: genetic_analysis::{Pallet, Call, Storage, Event<T>},
 		GeneticAnalysisOrders: genetic_analysis_orders::{Pallet, Call, Storage, Config<T>, Event<T>},
-		UserProfile: user_profile::{Pallet, Call, Storage, Event<T>},
-		Timestamp: pallet_timestamp::{Pallet, Call, Storage, Inherent},
-		RandomnessCollectiveFlip: pallet_randomness_collective_flip::{Pallet, Storage},
+		UserProfile: user_profile::{Pallet, Call, Storage, Event<T>}
 	}
 );
 
 parameter_types! {
 	pub const BlockHashCount: u64 = 250;
 	pub const SS58Prefix: u8 = 42;
+	pub BlockWeights: frame_system::limits::BlockWeights =
+	frame_system::limits::BlockWeights::simple_max(1024);
 }
 
-impl frame_system::Config for Test {
+impl system::Config for Test {
 	type BaseCallFilter = frame_support::traits::Everything;
 	type BlockWeights = ();
 	type BlockLength = ();
@@ -69,6 +75,7 @@ pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
 
 parameter_types! {
 	pub const MinimumPeriod: Moment = SLOT_DURATION / 2;
+	pub const GeneticAnalystPalletId: PalletId = PalletId(*b"dbio/gen");
 }
 
 impl pallet_timestamp::Config for Test {
@@ -77,35 +84,6 @@ impl pallet_timestamp::Config for Test {
 	type OnTimestampSet = ();
 	type MinimumPeriod = MinimumPeriod;
 	type WeightInfo = ();
-}
-
-impl pallet_randomness_collective_flip::Config for Test {}
-
-type Balance = u64;
-
-parameter_types! {
-	pub static ExistentialDeposit: Balance = 0;
-	pub const GeneticAnalystPalletId: PalletId = PalletId(*b"dbio/gen");
-	pub const GeneticAnalysisOrdersEscrowPalletId: PalletId = PalletId(*b"dbio/esc");
-}
-
-impl pallet_balances::Config for Test {
-	type MaxLocks = ();
-	type MaxReserves = ();
-	type ReserveIdentifier = [u8; 8];
-	/// The type for recording an account's balance.
-	type Balance = Balance;
-	/// The ubiquitous event type.
-	type Event = Event;
-	type DustRemoval = ();
-	type ExistentialDeposit = ExistentialDeposit;
-	type AccountStore = System;
-	type WeightInfo = ();
-}
-
-impl genetic_data::Config for Test {
-	type Event = Event;
-	type GeneticDataWeightInfo = ();
 }
 
 impl genetic_analysts::Config for Test {
@@ -133,13 +111,6 @@ impl genetic_analyst_qualifications::Config for Test {
 	type WeightInfo = ();
 }
 
-impl genetic_analysis::Config for Test {
-	type Event = Event;
-	type RandomnessSource = RandomnessCollectiveFlip;
-	type GeneticAnalysisOrders = GeneticAnalysisOrders;
-	type GeneticAnalysisWeightInfo = ();
-}
-
 impl genetic_analysis_orders::Config for Test {
 	type Event = Event;
 	type Currency = Balances;
@@ -151,33 +122,44 @@ impl genetic_analysis_orders::Config for Test {
 	type PalletId = GeneticAnalysisOrdersEscrowPalletId;
 }
 
-#[derive(Clone, Copy, PartialEq, Eq, Encode, Decode, Default, RuntimeDebug, TypeInfo)]
-pub struct EthereumAddress(pub [u8; 20]);
-
 impl user_profile::Config for Test {
 	type Event = Event;
 	type EthereumAddress = EthereumAddress;
 	type WeightInfo = ();
 }
 
-#[cfg(test)]
-use sp_io::TestExternalities;
+type Balance = u128;
 
-#[cfg(test)]
-pub struct ExternalityBuilder {
-	existential_deposit: u64,
+parameter_types! {
+	pub static ExistentialDeposit: Balance = 0;
 }
 
-#[cfg(test)]
+impl pallet_balances::Config for Test {
+	type MaxLocks = ();
+	type MaxReserves = ();
+	type ReserveIdentifier = [u8; 8];
+	/// The type for recording an account's balance.
+	type Balance = Balance;
+	/// The ubiquitous event type.
+	type Event = Event;
+	type DustRemoval = ();
+	type ExistentialDeposit = ExistentialDeposit;
+	type AccountStore = System;
+	type WeightInfo = ();
+}
+
+pub struct ExternalityBuilder {
+	existential_deposit: u128,
+}
+
 impl Default for ExternalityBuilder {
 	fn default() -> Self {
 		Self { existential_deposit: 1 }
 	}
 }
 
-#[cfg(test)]
 impl ExternalityBuilder {
-	pub fn existential_deposit(mut self, existential_deposit: u64) -> Self {
+	pub fn existential_deposit(mut self, existential_deposit: u128) -> Self {
 		self.existential_deposit = existential_deposit;
 		self
 	}
@@ -186,7 +168,7 @@ impl ExternalityBuilder {
 	}
 	pub fn build(&self) -> TestExternalities {
 		self.set_associated_consts();
-		let mut storage = frame_system::GenesisConfig::default().build_storage::<Test>().unwrap();
+		let mut storage = system::GenesisConfig::default().build_storage::<Test>().unwrap();
 		pallet_balances::GenesisConfig::<Test> { balances: { vec![] } }
 			.assimilate_storage(&mut storage)
 			.unwrap();
